% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{doctormoreau}
\alias{doctormoreau}
\title{"The Island of Doctor Moreau"}
\format{
A character vector with 4895 elements
}
\source{
\url{http://www.gutenberg.org/cache/epub/159/pg159.txt}
}
\usage{
doctormoreau
}
\description{
A data set containing the complete text of H. G. Wells 1896
novel "doctormoreau".
}
\keyword{datasets}
