% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{hgwells}
\alias{hgwells}
\alias{hgwells-package}
\title{Selected H. G. Wells Novels}
\description{
This package contains complete text of selected novels of
H. G. Wells formatted to be convenient for text analysis.
}
