% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{waroftheworlds}
\alias{waroftheworlds}
\title{"The War of the Worlds"}
\format{
A character vector with 6456 elements
}
\source{
\url{https://www.gutenberg.org/files/36/36-0.txt}
}
\usage{
waroftheworlds
}
\description{
A data set containing the complete text of H. G. Wells 1898
novel "waroftheworlds".
}
\keyword{datasets}
